Update program.py
Completed till ETL and completed the query for getting the nearest neighbor for a given (x, y) but the latitude and longitude in the given data set are not within the range (-180,180) so have to think how to normalize the data. Next I am going to start with the front end part.